refactor(ui): le composant metaux-mineraux-metropole est en tsx
