+++ packages/ui/src/components/statistiques/mineraux-metaux-metropole.module.css
@@ -1,0 +1,6 @@
+.grid-container {
+  display: grid;
+  grid-template-columns: 3fr 1fr;
+  grid-template-rows: repeat(4, auto);
+  gap: 3rem;
+}
