--- packages/ui/src/components/statistiques/mineraux-metaux-metropole.vue
@@ -1,13 +1,0 @@
-<template>
-  <PureMinerauxMetauxMetropole :getStats="getStats" />
-</template>
-
-<script setup lang="ts">
-import { fetchWithJson } from '@/api/client-rest'
-import { StatistiquesMinerauxMetauxMetropole } from 'camino-common/src/statistiques'
-import { CaminoRestRoutes } from 'camino-common/src/rest'
-import PureMinerauxMetauxMetropole from './pure-mineraux-metaux-metropole.vue'
-
-const getStats = async (): Promise<StatistiquesMinerauxMetauxMetropole> =>
-  fetchWithJson(CaminoRestRoutes.statistiquesMinerauxMetauxMetropole, {})
-</script>
